updaMeter

+++ Domain/MotorDto.cs
@@ -4,13 +4,13 @@ public class MotorDto
 {
     public string Name { get; set; }
     public int Speed { get; set; }
-    public int Meter { get; set; }
+    public float Meter { get; set; }
     public bool State { get; set; }
 }
 public class MotorDtoUpdate
 {
     public string Name { get; set; }
     public int? Speed { get; set; }
-    public int? Meter { get; set; }
+    public float? Meter { get; set; }
     public bool? State { get; set; }
 }
